utility for selects
